IndexBased: split funcs not specific to algorithm to separate file
