+++ benchmarks/SCAN/IndexBased/BUILD
@@ -11,6 +11,7 @@ cc_library(
     ],
     visibility = ["//visibility:public"],
     deps = [
+        ":utils",
         "//benchmarks/Connectivity/UnionFind:union_find_rules",
         "//benchmarks/TriangleCounting/ShunTangwongsan15:Triangle_main",
         "//ligra:bridge",
@@ -25,6 +26,21 @@ cc_library(
         "//pbbslib:sample_sort",
         "//pbbslib:seq",
         "//pbbslib:utilities",
+    ],
+)
+
+cc_library(
+    name = "utils",
+    srcs = [
+        "utils.cc",
+    ],
+    hdrs = [
+        "utils.h",
+    ],
+    deps = [
+        "//ligra:graph",
+        "//ligra:macros",
+        "//pbbslib:seq",
     ],
 )
