+++ benchmarks/SCAN/IndexBased/scan.cc
@@ -1,12 +1,10 @@
 #include "benchmarks/SCAN/IndexBased/scan.h"
 
 #include <algorithm>
-#include <sstream>
 #include <tuple>
 #include <utility>
 
 #include "benchmarks/Connectivity/UnionFind/union_find_rules.h"
-#include "ligra/bridge.h"
 #include "ligra/pbbslib/sparse_table.h"
 #include "pbbslib/parallel.h"
 
@@ -120,18 +118,6 @@ void AttachNoncoresToClusters(
 
 }  // namespace
 
-std::ostream& operator<<(std::ostream& os, UnclusteredType unclustered_type) {
-  switch (unclustered_type) {
-    case UnclusteredType::kHub:
-      os << "hub";
-      break;
-    case UnclusteredType::kOutlier:
-      os << "outlier";
-      break;
-  }
-  return os;
-}
-
 Clustering Index::Cluster(const uint64_t mu, const float epsilon) const {
   const pbbs::sequence<uintE> cores{core_order_.GetCores(mu, epsilon)};
   if (cores.empty()) {
@@ -159,37 +145,6 @@ Clustering Index::Cluster(const uint64_t mu, const float epsilon) const {
   AttachNoncoresToClusters(
       neighbor_order_, cores, core_similar_edge_counts, &clustering);
   return clustering;
-}
-
-size_t CleanClustering(Clustering* clustering) {
-  const size_t num_vertices{clustering->size()};
-  pbbs::sequence<uintE> cluster_relabel_map(num_vertices, 0U);
-  par_for(0, num_vertices, [&](const size_t i) {
-    const uintE cluster_id{(*clustering)[i]};
-    if (cluster_id != kUnclustered && cluster_relabel_map[cluster_id] == 0) {
-      cluster_relabel_map[cluster_id] = 1;
-    }
-  });
-  const size_t num_clusters{pbbslib::scan_add_inplace(cluster_relabel_map)};
-  par_for(0, num_vertices, [&](const size_t i) {
-    const uintE cluster_id{(*clustering)[i]};
-    if (cluster_id != kUnclustered) {
-      (*clustering)[i] = cluster_relabel_map[cluster_id];
-    }
-  });
-  return num_clusters;
-}
-
-std::string ClusteringToString(const Clustering& clustering) {
-  std::ostringstream str;
-  str << "{";
-  for (size_t i = 0; i < clustering.size(); i++) {
-    str << ' ' << i << ':';
-    str << (clustering[i] == kUnclustered
-            ? "n/a" : std::to_string(clustering[i]));
-  }
-  str << " }";
-  return str.str();
 }
 
 }  // namespace indexed_scan
